requirements.txt
biopython==1.78
numpy==1.19.5
torch==1.1.0
